Added health system and started factions code
The health script had everything needed to notify other scripts of different states and handle tracking the entity health aswell as handling entity destruction.

The factions enum will be used on entities to denote what faction they are a part of so the AI can use that to chase or not chase that entity.