Some nice worm and food updates
Worms can now die.

When a worm dies it drops food along where all it's body segments existed.

Food spawning has been moved to a dedicated food spawner so all the food doesn't take up the hierarchy and is under 1 parent object.

When the player worm specifically dies the game is over and a player death event is broadcasted.

When the game is over a "YOU HAVE DIED" UI is shown to the player to indicate game over.